Add NextJS template.
This builds out a NextJS based template. We don't need sonar or
a reverse proxy since this is a single container.

I left in the code for using GPUs. Maybe someday we'll do GPU
work using NodeJS. That *might* be better than Python. Who knows.
Crazier things have happened.

This uses Shellac instead of full-fledged Varnish as there's a few
bugs which cause conflicts when trying to use Varnish and NextJS
that I wasn't able to resolve:

- https://github.com/allenai/varnish-mui/issues/141
- https://github.com/allenai/varnish-mui/issues/142

When those are fixed I expect someone to circle back and add Varnish
and other goodies.